2287 - move locust data creation into management cmd
